Refactor exception
